Refactor init
