+++ scala/ppml/src/main/scala/com/intel/analytics/bigdl/ppml/attestation/VerificationCLI.scala
@@ -22,6 +22,9 @@ import scopt.OptionParser
 
 import java.util.Base64
 
+import com.intel.analytics.bigdl.ppml.attestation.service._ 
+import com.intel.analytics.bigdl.ppml.attestation.verifier._ 
+
 /**
  * Simple Command Line tool to verify attestation service
  */
