+++ scala/ppml/src/main/scala/com/intel/analytics/bigdl/ppml/attestation/generator/QuoteGeneratorCmd.scala
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.bigdl.ppml.attestation
+package com.intel.analytics.bigdl.ppml.attestation.generator
 import org.apache.logging.log4j.LogManager
 import java.io.{File, FileInputStream, FileOutputStream}
 import scopt.OptionParser
+
+import com.intel.analytics.bigdl.ppml.attestation._
 
 object QuoteGeneratorCmd {
     def main(args: Array[String]): Unit = {
