+++ scala/ppml/src/main/scala/com/intel/analytics/bigdl/ppml/attestation/service/AttestationService.scala
