+++ scala/ppml/src/main/scala/com/intel/analytics/bigdl/ppml/attestation/utils/RegisterMrEnclave.scala
