+++ scala/ppml/src/main/scala/com/intel/analytics/bigdl/ppml/attestation/verifier/SGXQuoteVerifierImpl.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.bigdl.ppml.attestation
+package com.intel.analytics.bigdl.ppml.attestation.verifier
 
 /**
  * Verify SGX quote with SGX SDK
