+++ scala/ppml/src/test/scala/com/intel/analytics/bigdl/ppml/attestation/verifier/SGXDCAPQuoteVerifierImplSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.bigdl.ppml.attestation
+package com.intel.analytics.bigdl.ppml.attestation.verifier
 
 import com.intel.analytics.bigdl.dllib.common.zooUtils
 import java.io.{BufferedOutputStream, BufferedInputStream};
